+++ api/src/main/java/com/service/ProviderService.java
@@ -1,0 +1,9 @@
+package com.service;
+
+public interface ProviderService {
+    /**
+     * 获取服务端参数*
+     * @return
+     */
+    String getProviderInfo();
+}
