+++ consumer/src/main/java/com/fallback/ErrorServiceFallback.java
@@ -1,0 +1,12 @@
+package com.fallback;
+
+import com.service.ErrorService;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ErrorServiceFallback implements ErrorService {
+    @Override
+    public String errorInfo() {
+        return "error 快速失败";
+    }
+}
