+++ consumer/src/main/java/com/service/ErrorService.java
@@ -1,0 +1,14 @@
+package com.service;
+
+import com.fallback.ErrorServiceFallback;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@FeignClient(name = "errorProvider", path = "", fallback = ErrorServiceFallback.class)
+public interface ErrorService {
+
+    @ResponseBody
+    @GetMapping("/errorInfo")
+    public String errorInfo();
+}
